fixpix-ai-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
requests==2.31.0
structlog==24.1.0
python-dotenv==1.0.1
# AI / ML
numpy==1.26.3
opencv-python-headless==4.9.0.80
torch==2.1.2
torchvision==0.16.2
facexlib==0.3.0
gfpgan==1.3.8
codeformer-pip
realesrgan
rembg
segment-anything
# Generative AI
diffusers
transformers
accelerate
controlnet-aux
# Utilities
aiofiles==23.2.1
uuid
Pillow
xgboost
